requirements.txt
alembic==1.4.2
appdirs==1.4.4
astroid==2.4.2
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.3
distlib==0.3.1
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.1
filelock==3.0.12
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gunicorn==19.9.0
idna==2.8
importlib-metadata==1.7.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.5
pylint==2.5.3
pyperclip==1.8.0
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
requests==2.22.0
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.3
virtualenv==20.0.25
visitor==0.1.3
Werkzeug==0.15.6
wrapt==1.12.1
WTForms==2.3.1
zipp==3.1.0
